lock work?

+++ Blog/blog_delete.php
@@ -10,13 +10,27 @@
 
         $suID = 0;
         $eID = 0;
+        $fID = $uID;
 
         if(!empty($_GET['eID'])){
             $eID = $_GET['eID'];
         }
         if(!empty($_GET['duID'])){
             $duID = $_GET['duID'];
+            $fID = $_GET['duID'];
         }
+
+
+        $stmt = $conn->prepare("SELECT locked FROM user WHERE ID = ?");  //gets if the user is locked
+        $stmt->bind_param("i", $fID);   
+        $stmt->execute(); 
+        $result = $stmt->get_result(); 
+
+        if ($result->num_rows > 0) {
+            while($row = $result->fetch_assoc()) {    
+                $lock= $row["locked"];
+            }
+        } 
         
 
 
@@ -27,33 +41,46 @@
     ----------------------------------------*/
 
         if ($res["userType"] == "endUser"){
-            if ($eID != 0){
-                $stmt = $conn->prepare("DELETE FROM blog_entry WHERE blog_entry.ID = ? ");  // deletes entries by specific id
-                $stmt->bind_param("i", $eID); 
-                $stmt->execute();  
-                $data = ["Action"=>"Entry deleted"];
-                jsonWrite($version,$data);
+            if ($lock == 0){
+                if ($eID != 0){    
+                    $stmt = $conn->prepare("DELETE FROM blog_entry WHERE blog_entry.ID = ? ");  // deletes entries by specific id
+                    $stmt->bind_param("i", $eID); 
+                    $stmt->execute();  
+                    $data = ["Action"=>"Entry deleted"];
+                    jsonWrite($version,$data);
+                }    
+                else{
+                    errorWrite($version,"Wrong inputs");
+                }
             }
             else{
-                errorWrite($version,"Wrong inputs");
+                $data = ["Blog"=>"you are locked"];
+                jsonWrite($version,$data); 
             }
         }
         else if ($res["userType"] == "admin"){
-            if ($duID != 0){
-                $stmt = $conn->prepare("DELETE FROM blog WHERE blog.ID = ?"); // deletes blogs by specific id
-                $stmt->bind_param("i", $duID); 
-                $stmt->execute(); 
+            if ($lock == 0){
+                if ($duID != 0){    
+                    $stmt = $conn->prepare("DELETE FROM blog WHERE blog.ID = ?"); // deletes blogs by specific id
+                    $stmt->bind_param("i", $duID); 
+                    $stmt->execute(); 
 
-                $stmt = $conn->prepare("DELETE FROM blog_entry WHERE blog_entry.uID = ?"); // deletes entries from the blog when deleted
-                $stmt->bind_param("i", $duID); 
-                $stmt->execute(); 
-                
-                $data = ["Action"=>"Blog deleted"];
-                jsonWrite($version,$data);
+                    $stmt = $conn->prepare("DELETE FROM blog_entry WHERE blog_entry.uID = ?"); // deletes entries from the blog when deleted
+                    $stmt->bind_param("i", $duID); 
+                    $stmt->execute(); 
+                    
+                    $data = ["Action"=>"Blog deleted"];
+                    jsonWrite($version,$data);   
+                }
+                else{
+                    errorWrite($version,"Wrong inputs");
+                }
             }
             else{
-                errorWrite($version,"Wrong inputs");
+                $data = ["Blog"=>"user is locked"];
+                jsonWrite($version,$data); 
             }
+            
         }
         else{
             errorWrite($version,"No user");
